# apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Atom
  version: 0.1.0
  description: |
    Identity and Authorization service for IoT and cloud-native systems.

    All endpoints except `GET /health`, `POST /auth/login`, and
    `GET /.well-known/jwks.json` require a Bearer token:

    ```
    Authorization: Bearer <token>
    ```

    Two token types are accepted:
    - **JWT** (`eyJ...`) — returned by `/auth/login`, short-lived (default 1 h)
    - **API key** (`atom_...`) — long-lived, created per entity

    Endpoints marked **RequireManage** additionally require the caller to hold an
    `allow` + `scope=all` policy binding covering the `manage` capability.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Reusable components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT (`eyJ...`) or API key (`atom_<id>_<secret>`)

  # ── Schemas ──────────────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: not found

    # ── Auth ───────────────────────────────────────────────────────────────
...
    TenantStatus:
      type: string
      enum: [active, inactive, frozen, deleted]
      default: active
    Tenant:
      type: object
      required: [id, name, status, tags, attributes, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        route:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TenantStatus'
        tags:
          type: array
          items:
            type: string
        attributes:
          type: object
          description: Arbitrary JSONB metadata; reusable for Magistrala domain metadata.
        created_by:
          type: string
          format: uuid
          nullable: true
        updated_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTenant:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: factory-1
        route:
          type: string
        tags:
          type: array
          items:
            type: string
        attributes:
          type: object
    UpdateTenant:
      type: object
      properties:
        name:
          type: string
        route:
          type: string
        tags:
          type: array
          items:
            type: string
        attributes:
          type: object
          description: Replaces all existing attributes when provided
    TenantList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Tenant'
        total:
          type: integer

  # ── Reusable responses ───────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: missing Authorization header
    Forbidden:
      description: Caller lacks `manage` scope=all
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not found
    Conflict:
      description: Already exists (unique constraint)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: already exists

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /tenants:
    get:
      operationId: listTenants
      summary: List tenants
      tags: [Tenants]
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: route
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TenantStatus'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createTenant
      summary: Create a tenant
      description: "**RequireManage.** A tenant is an isolation boundary, not a principal."
      tags: [Tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenant'
            example:
              name: factory-1
              route: factory-1
              tags: [pilot]
              attributes:
                magistrala:
                  region: eu
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
  /tenants/{id}:
    get:
      operationId: getTenant
      summary: Get a tenant
      tags: [Tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateTenant
      summary: Update a tenant
      description: "**RequireManage.**"
      tags: [Tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenant'
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteTenant
      summary: Soft-delete a tenant (status = deleted)
      description: "**RequireManage.** Sets status = `deleted`; the row is retained."
      tags: [Tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...